Remove old header
